pyproject.toml
[tool.poetry]
name = "hw14"
version = "0.1.0"
description = ""
authors = ["andr"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiosmtplib = "2.0.2"
alabaster = "0.7.16"
alembic = "1.13.1"
annotated-types = "0.7.0"
anyio = "4.4.0"
async-timeout = "4.0.3"
babel = "2.15.0"
bcrypt = "4.1.3"
blinker = "1.8.2"
certifi = "2024.6.2"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
cloudinary = "1.40.0"
cryptography = "42.0.8"
dnspython = "2.6.1"
docutils = "0.21.2"
ecdsa = "0.19.0"
email-validator = "2.1.1"
exceptiongroup = "1.2.1"
fastapi = "0.111.0"
fastapi-cli = "0.0.4"
fastapi-limiter = "0.1.6"
fastapi-mail = "1.4.1"
greenlet = "3.0.3"
h11 = "0.14.0"
httpcore = "1.0.5"
httptools = "0.6.1"
httpx = "0.27.0"
idna = "3.7"
imagesize = "1.4.1"
iniconfig = "2.0.0"
jinja2 = "3.1.4"
libgravatar = "1.0.4"
mako = "1.3.5"
markdown-it-py = "3.0.0"
markupsafe = "2.1.5"
mdurl = "0.1.2"
orjson = "3.10.3"
packaging = "24.0"
passlib = "1.7.4"
pluggy = "1.5.0"
psycopg2-binary = "2.9.9"
pyasn1 = "0.6.0"
pycparser = "2.22"
pydantic = "2.7.3"
pydantic-settings = "2.3.1"
pydantic-core = "2.18.4"
pygments = "2.18.0"
pytest = "8.2.2"
pytest-mock = "3.14.0"
python-dotenv = "1.0.1"
python-jose = "3.3.0"
python-multipart = "0.0.9"
pyyaml = "6.0.1"
redis = "5.1.0b6"
requests = "2.32.3"
rich = "13.7.1"
rsa = "4.9"
shellingham = "1.5.4"
six = "1.16.0"
sniffio = "1.3.1"
snowballstemmer = "2.2.0"
sphinx = "7.3.7"
sphinxcontrib-applehelp = "1.0.8"
sphinxcontrib-devhelp = "1.0.6"
sphinxcontrib-htmlhelp = "2.0.5"
sphinxcontrib-jsmath = "1.0.1"
sphinxcontrib-qthelp = "1.0.7"
sphinxcontrib-serializinghtml = "1.1.10"
sqlalchemy = "2.0.30"
starlette = "0.37.2"
tomli = "2.0.1"
typer = "0.12.3"
typing-extensions = "4.12.1"
ujson = "5.10.0"
urllib3 = "2.2.1"
uvicorn = "0.30.1"
uvloop = "0.19.0"
watchfiles = "0.22.0"
websockets = "12.0"
aiosqlite = "^0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
